Fixed asin and atan functions
It was mostly just a problem with arg and those darn coterminal angles and whatnot and oh boy you don't wanna know how long it took me to figure that out. Also I think we're all done except for some equality checkers that we need to have say "sorry bro but ya just can't use > < >= <= bc that's not how complex numbers work" or smth close to that.